+++ changelog.md
@@ -1,7 +1,12 @@
 ### Additions
 - Added a sound to Otherworldly Energy infusing
 - Added some particles when an item is infused by Otherworldly Energy
+- Added Otherworldly Stone Bricks
 
 ### Changes
 - Shatter Coils will no longer attack tamed animals
 - Otherworldly Energy now applies a pickup delay to the result
+- Changed colors on Otherworldly Stone
+
+### Fixes
+- Fixed Runic Chisel being odd in offhand